+++ exporter/exporterhelper/common.go
@@ -23,7 +23,6 @@ import (
 	"go.opentelemetry.io/collector/component"
 	"go.opentelemetry.io/collector/component/componenthelper"
 	"go.opentelemetry.io/collector/config/configmodels"
-	"go.opentelemetry.io/collector/consumer/consumererror"
 )
 
 // ComponentSettings for timeout. The timeout applies to individual attempts to send data to the backend.
@@ -46,8 +45,9 @@ type request interface {
 	// setContext updates the Context of the requests.
 	setContext(context.Context)
 	export(ctx context.Context) error
-	// Returns a new request that contains the items left to be sent.
-	onPartialError(consumererror.PartialError) request
+	// Returns a new request may contain the items left to be sent if some items failed to process and can be retried.
+	// Otherwise, it should return the original request.
+	onError(error) request
 	// Returns the count of spans/metric points or log records.
 	count() int
 }