+++ receiver/scraperhelper/scrapercontroller_test.go
@@ -327,7 +327,7 @@ func getExpectedShutdownErr(test metricsTestCase) error {
 		}
 	}
 
-	return consumererror.CombineErrors(errs)
+	return consumererror.Combine(errs)
 }
 
 func assertChannelsCalled(t *testing.T, chs []chan bool, message string) {