+++ service/internal/builder/extensions_builder.go
@@ -52,7 +52,7 @@ func (exts Extensions) StartAll(ctx context.Context, host component.Host) error 
 	for _, ext := range exts {
 		ext.logger.Info("Extension is starting...")
 
-		if err := ext.Start(ctx, host); err != nil {
+		if err := ext.Start(ctx, newHostWrapper(host, ext.logger)); err != nil {
 			return err
 		}
 
@@ -71,7 +71,7 @@ func (exts Extensions) ShutdownAll(ctx context.Context) error {
 		}
 	}
 
-	return consumererror.CombineErrors(errs)
+	return consumererror.Combine(errs)
 }
 
 func (exts Extensions) NotifyPipelineReady() error {
@@ -99,7 +99,7 @@ func (exts Extensions) NotifyPipelineNotReady() error {
 		}
 	}
 
-	return consumererror.CombineErrors(errs)
+	return consumererror.Combine(errs)
 }
 
 func (exts Extensions) ToMap() map[configmodels.NamedEntity]component.Extension {